# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轨道追踪器后端 API 服务，提供卫星数据查询、用户认证、收藏管理等功能",
        "title": "Orbital Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@orbital-tracker.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/satellites": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卫星"
                ],
                "summary": "获取卫星列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/satellites/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卫星"
                ],
                "summary": "搜索卫星",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/satellites/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卫星"
                ],
                "summary": "获取卫星详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NORAD ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/satellites/{id}/tle": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "卫星"
                ],
                "summary": "获取卫星TLE数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NORAD ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/password": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "密码信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_dto_request.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/profile": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户资料",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户资料",
                "parameters": [
                    {
                        "description": "更新信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_dto_request.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/settings": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户设置",
                "parameters": [
                    {
                        "description": "设置信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_dto_request.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SituationBak_internal_pkg_utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "SituationBak_internal_dto_request.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "SituationBak_internal_dto_request.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "maxLength": 255
                },
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "SituationBak_internal_dto_request.UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string",
                    "enum": [
                        "zh-CN",
                        "en-US"
                    ]
                },
                "satellite_limit": {
                    "type": "integer",
                    "maximum": 50000,
                    "minimum": 100
                },
                "show_debris": {
                    "type": "boolean"
                },
                "theme": {
                    "type": "string",
                    "enum": [
                        "dark",
                        "light",
                        "auto"
                    ]
                }
            }
        },
        "SituationBak_internal_pkg_utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT 认证令牌，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}